ssr prepass
