+++ examples/.test/diagnostics-big-router/src/utils/trpc.ts
@@ -30,5 +30,4 @@ export const trpc = createTRPCNext<AppRouter>({
       ],
     };
   },
-  ssr: true,
 });
