+++ examples/next-edge-runtime/src/utils/trpc.ts
@@ -28,5 +28,4 @@ export const trpc = createTRPCNext<AppRouter>({
       ],
     };
   },
-  ssr: true,
 });
