+++ examples/next-minimal-starter/src/utils/trpc.ts
@@ -1,5 +1,6 @@
 import { httpBatchLink } from '@trpc/client';
 import { createTRPCNext } from '@trpc/next';
+import { ssrPrepass } from '@trpc/next/pages/ssr';
 import type { AppRouter } from '../pages/api/trpc/[trpc]';
 
 function getBaseUrl() {
@@ -29,4 +30,5 @@ export const trpc = createTRPCNext<AppRouter>({
     };
   },
   ssr: true,
+  ssrPrepass,
 });
