+++ packages/next/README.md
@@ -55,7 +55,7 @@ export const trpc = createTRPCNext<AppRouter>({
       ],
     };
   },
-  ssr: true,
+  ssr: false,
 });
 ```
