+++ packages/next/turbo.json
@@ -2,6 +2,8 @@
   "$schema": "https://turborepo.org/schema.json",
   "extends": ["//"],
   "pipeline": {
-    "codegen-entrypoints": { "outputs": ["package.json", "app-dir/**"] }
+    "codegen-entrypoints": {
+      "outputs": ["package.json", "app-dir/**", "pages/**"]
+    }
   }
 }
