+++ www/docs/client/nextjs/ssr.md
@@ -21,10 +21,13 @@ Additionally, consider [`Response Caching`](../../server/caching.md).
 ```tsx title='utils/trpc.ts'
 import { httpBatchLink } from '@trpc/client';
 import { createTRPCNext } from '@trpc/next';
+import { ssrPrepass } from '@trpc/next/pages/ssr';
 import superjson from 'superjson';
 import type { AppRouter } from './api/trpc/[trpc]';
 
 export const trpc = createTRPCNext<AppRouter>({
+  ssr: true,
+  ssrPrepass,
   config(opts) {
     const { ctx } = opts;
     if (typeof window !== 'undefined') {
@@ -61,7 +64,6 @@ export const trpc = createTRPCNext<AppRouter>({
       ],
     };
   },
-  ssr: true,
 });
 ```
